Gppwtents

+++ trefftz/scalarmappedfe.hpp
@@ -171,7 +171,7 @@ namespace ngfem
     // public:
     //	NGS_DLL_HEADER virtual std::list<std::tuple<std::string,double>> Timing
     //() const;
-    virtual float GetWavespeed () const
+    virtual double GetWavespeed () const
     {
       return 0;
     } // ugly parent hack for trefftzwave
